splitting common.py (part.2): moved remains to other modules, common.py is no more
